+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * [EACQAPW-5528] PaymentSystem Tests
 * [EACQAPW-5532] TDSCertsManager Tests
 * [EACQAPW-5551] Move to Static linking
+* [EACQAPW-5556] CardPaymentProcess Tests
 
 ### Fixed